Refactor pcap code, split in different files
